v0.7.2
[Enhancement] Fixed issues in the code which caused warnings in MSVC and clang compilers.

[Enhancement] Fixed errors in `get_heuristics_list` where for certain heuristics mode in older cuDNN versions, the heuristics list might be incorrect.

[Bug fixes] Fixed several test cases failing on unsupported GPUs to exit gracefully.

[Samples] Added a sample to showcase fp8 convolution forward in Nvidia Hopper GPUs. The sample also showcases post convolution book-keeping operations such as scaling and absolute maximum reduction.

[Samples] Added a sample which converts fp16 tensor to fp8 and performs transpose and absolute maximum reduction.

[Samples] Added a sample to demonstrate Max pooling operation including tensor index dump, necessary to speed up the backward pass.

[Samples] Added a sample to showcase the backward pooling operation.

+++ samples/cpu_references.h
@@ -23,6 +23,7 @@
 #pragma once
 
 #include "helpers.h"
+#include <cmath>
 
 template <typename T_ELEM>
 void weightGrad_cpu_ref(
@@ -60,22 +61,22 @@ void weightGrad_cpu_ref(
     bool isConv = true;  //(CUDNN_CONVOLUTION == mode) ;
 
     // For every filter pixel (k x c x r x s)
-    for (int ci = 0; ci < inDims[1]; ci++) {               // Loop over filter output pixels
-        for (int ri = 0; ri < filDims[2]; ri++) {          //        ^
-            for (int si = 0; si < filDims[3]; si++) {      //    ^
-                for (int ki = 0; ki < filDims[0]; ki++) {  // ^
-                    int filIdx = ki * filterStride[0] + ci * filterStride[1] + ri * filterStride[2] + si * filterStride[3];
+    for (size_t ci = 0; ci < (size_t)inDims[1]; ci++) {               // Loop over filter output pixels
+        for (size_t ri = 0; ri < (size_t)filDims[2]; ri++) {          //        ^
+            for (size_t si = 0; si < (size_t)filDims[3]; si++) {      //    ^
+                for (size_t ki = 0; ki < (size_t)filDims[0]; ki++) {  // ^
+                    int64_t filIdx = ki * filterStride[0] + ci * filterStride[1] + ri * filterStride[2] + si * filterStride[3];
                     float val = 0.f;
                     // For every image (n)
                     for (int ni = 0; ni < inDims[0]; ni++) {  // Sum over the batch
-                        int offset_image = ni * inStride[0] + ci * inStride[1];
-                        int offset_diff  = ni * diffStride[0] + ki * diffStride[1];
+                        int64_t offset_image = ni * inStride[0] + ci * inStride[1];
+                        int64_t offset_diff  = ni * diffStride[0] + ki * diffStride[1];
                         // For every pixel in diff (p x q)
                         for (int pi = 0; pi < diffDims[2]; pi++) {      // Sum over the pixels of diff
                             for (int qi = 0; qi < diffDims[3]; qi++) {  //  ^
                                 // Fetch the value in image and diff, product and accumulate
-                                int y = pi * stride[0] - pad[0];
-                                int x = qi * stride[1] - pad[1];
+                                int64_t y = pi * stride[0] - pad[0];
+                                int64_t x = qi * stride[1] - pad[1];
                                 // Convolution = Correlation with a flipped filter
                                 // So basically, for the convolution, we replace r by dim-1-r 
                                 // and s by dim-1-s to "flip" the filter.
@@ -92,11 +93,11 @@ void weightGrad_cpu_ref(
                                     x += si * dilation[1];
                                 }
                                 // Image value
-                                int inBounds = ((x >= 0) && (x < inDims[3]) && (y >= 0) && (y < inDims[2]));
+                                int64_t inBounds = ((x >= 0) && (x < inDims[3]) && (y >= 0) && (y < inDims[2]));
                                 if (inBounds) {
-                                    int imIdx = offset_image + y * inStride[2] + x * inStride[3];
+                                    int imIdx = static_cast<int>(offset_image + y * inStride[2] + x * inStride[3]);
                                     // Diff value
-                                    int diffIdx = offset_diff + pi * diffStride[2] + qi * diffStride[3];
+                                    int diffIdx = static_cast<int>(offset_diff + pi * diffStride[2] + qi * diffStride[3]);
                                     // Prod and accumulate
                                     T_ELEM imTmp   = image[imIdx];
                                     T_ELEM diffTmp = diffData[diffIdx];
@@ -159,7 +160,7 @@ void gen_stats_cpu(
     const int64_t inputSize,
     const int64_t* inputDims)
 {
-    std::vector<int> totals(inputDims[0], 0);
+    std::vector<int64_t> totals((size_t)inputDims[0], 0);
     for (int i = 0; i < inputSize; i++) {
         int channel_index = i % inputDims[0];
 
@@ -169,7 +170,7 @@ void gen_stats_cpu(
     }
 
     // Calculate the mean for each channel. Assumes NHWC format.
-    for (int i = 0; i < outputData.size(); i++) {
+    for (size_t i = 0; i < outputData.size(); i++) {
         outputData[i].first = outputData[i].first / totals[i];
     }
 
@@ -182,7 +183,7 @@ void gen_stats_cpu(
     }
 
     // Calculate the variance for the channel. Assumes NHWC format.
-    for (int i = 0; i < outputData.size(); i++) {
+    for (size_t i = 0; i < outputData.size(); i++) {
         outputData[i].second = outputData[i].second / totals[i];
     }
 }
@@ -220,7 +221,7 @@ void conv_cpu_ref(
     const int64_t* dilation,
     int64_t nbDims) 
 {
-    int imDims = nbDims - 2;
+    int64_t imDims = nbDims - 2;
     float alpha     = 1.0f;
     float beta      = 0.0;
     // Some sanity checks
@@ -246,13 +247,13 @@ void conv_cpu_ref(
     bool isConv = true;  //(CUDNN_CONVOLUTION == mode) ;
 
     // Number of pixels in output
-    int nPixelsOut = 1;
+    int64_t nPixelsOut = 1;
     for (int i = 2; i < nbDims; i++) {
         nPixelsOut *= diffDims[i];
     }
 
     // Number of pixels in filter
-    int nPixelsFil = 1;
+    int64_t nPixelsFil = 1;
     for (int i = 2; i < nbDims; i++) {
         nPixelsFil *= filDims[i];
     }
@@ -267,7 +268,7 @@ void conv_cpu_ref(
     for (int64_t ni = 0; ni < diffDims[0]; ni++) {
         // For each outer feature layer of the output image
         for (int ki_outer = 0; ki_outer < diffDims[1] / resizeFactor; ki_outer++) {
-            int outputOffset = ni * diffStride[0] / resizeFactor + ki_outer * diffStride[1];
+            int64_t outputOffset = ni * diffStride[0] / resizeFactor + ki_outer * diffStride[1];
             // For every pixel in this output image's feature layer
             for (int outId = 0; outId < nPixelsOut; outId++) {
                 // Get output pixel ids
@@ -283,8 +284,8 @@ void conv_cpu_ref(
                     T_MATH tmp = 0;
                     // For each outer feature layer of the input image and filter
                     for (int ci = 0; ci < inDims[1] / resizeFactor; ci++) {
-                        int inputOffset = ni * inStride[0] / resizeFactor + ci * inStride[1];
-                        int filterOffset = (ki_outer * resizeFactor + ki_inner) * filStride[0] / resizeFactor + ci * filStride[1];
+                        int64_t inputOffset = ni * inStride[0] / resizeFactor + ci * inStride[1];
+                        int64_t filterOffset = (ki_outer * resizeFactor + ki_inner) * filStride[0] / resizeFactor + ci * filStride[1];
                         // Now for every pixel in the filter
                         for (int filId = 0; filId < nPixelsFil; filId++) {
                             // Get the position of the pixel
@@ -304,14 +305,14 @@ void conv_cpu_ref(
                                 inside &= (tmpIds[d] >= 0 && tmpIds[d] < inDims[2 + d]);
                             }
                             if (inside) {
-                                int actualTmpId = inputOffset + dim2lin(tmpIds, (inStride) + 2, imDims);
+                                int64_t actualTmpId = inputOffset + dim2lin(tmpIds, (inStride) + 2, imDims);
                                 // int actualFilId = filterOffset + filId ;
-                                int actualFilId = filterOffset + dim2lin(filIds, (filStride) + 2, imDims);
+                                int64_t actualFilId = filterOffset + dim2lin(filIds, (filStride) + 2, imDims);
 
                                 // For each inner feature layer of the input image and filter
                                 for (int i = 0; i < resizeFactor; i++) {
-                                    T_ELEM fval = filterData[actualFilId * resizeFactor + i];
-                                    T_ELEM ival = inputData[actualTmpId * resizeFactor + i];
+                                    T_ELEM fval = filterData[(size_t)(actualFilId * resizeFactor + i)];
+                                    T_ELEM ival = inputData[(size_t)(actualTmpId * resizeFactor + i)];
                                     tmp         = doFma(fval, ival, tmp);
                                 }
                             }
@@ -319,7 +320,7 @@ void conv_cpu_ref(
                     }
 
                     // Store final result in proper position in output image
-                    int actualOutId = outputOffset + dim2lin(outIds, (diffStride) + 2, imDims);
+                    int64_t actualOutId = outputOffset + dim2lin(outIds, (diffStride) + 2, imDims);
                     doEpilog(outputData, actualOutId * resizeFactor + ki_inner, alpha * tmp, beta);
                 }
             }
@@ -370,16 +371,16 @@ void dataGrad_cpu_ref(
         for (int ci = 0; ci < outDims[1]; ci++) {
             for (int hi = 0; hi < outDims[2]; hi++) {
                 for (int wi = 0; wi < outDims[3]; wi++) {
-                    int outIdx = ni * outStride[0] + ci * outStride[1] + hi * outStride[2] + wi * outStride[3];
+                    int64_t outIdx = ni * outStride[0] + ci * outStride[1] + hi * outStride[2] + wi * outStride[3];
                     float val  = 0.0;
 
                     // For every diff channel (k)
                     for (int ki = 0; ki < inDims[1]; ki++) {  // Sum over k channels
-                        int offset_filter = ki * filStride[0] + ci * filStride[1];
-                        int offset_diff   = ni * inStride[0] + ki * inStride[1];
+                        int64_t offset_filter = ki * filStride[0] + ci * filStride[1];
+                        int64_t offset_diff   = ni * inStride[0] + ki * inStride[1];
                         // For every pixel if filter (r x s)
                         for (int ri = 0; ri < filDims[2]; ri++) {
-                            int p = hi + pad[0];
+                            int64_t p = hi + pad[0];
 
                             if (isConv) {
                                 p -= (filDims[2] - 1 - ri) * dilation[0];
@@ -394,7 +395,7 @@ void dataGrad_cpu_ref(
                             p /= stride[0];
 
                             for (int si = 0; si < filDims[3]; si++) {
-                                int q = wi + pad[1];
+                                int64_t q = wi + pad[1];
 
                                 // Fetch the value in filter and diff, product and accumulate
                                 // So basically, for the convolution, we replace r by dim-1-r 
@@ -415,10 +416,10 @@ void dataGrad_cpu_ref(
 
                                 int inBounds = ((p >= 0) && (p < inDims[2]) && (q >= 0) && (q < inDims[3]));
                                 if (inBounds) {
-                                    int filterIdx = offset_filter + ri * filStride[2] + si * filStride[3];
-                                    int diffIdx   = offset_diff + p * inStride[2] + q * inStride[3];
-                                    T_ELEM imTmp  = top_diff[diffIdx];
-                                    T_ELEM filTmp = weight[filterIdx];
+                                    int64_t filterIdx = offset_filter + ri * filStride[2] + si * filStride[3];
+                                    int64_t diffIdx   = offset_diff + p * inStride[2] + q * inStride[3];
+                                    T_ELEM imTmp  = top_diff[(size_t)diffIdx];
+                                    T_ELEM filTmp = weight[(size_t)filterIdx];
                                     val           = doFma(filTmp, imTmp, val);
                                 }
                             }
